requirements.txt
fastapi
uvicorn
pypdf
sentence-transformers
faiss-cpu
numpy
tqdm
pickle-mixin
requests
pymupdf